Can calculate the total price of all the order

+++ ImageStream/Product.swift
@@ -14,7 +14,7 @@ class Product:  NSObject , NSCoding {
     var name : String?
     var productImage : String?
     var cellImage : String?
-    var price : Double?
+    var price : Int?
     
    override init() {
         super.init()
@@ -26,7 +26,7 @@ class Product:  NSObject , NSCoding {
         self.name = aDecoder.decodeObject(forKey:"name") as! String?
         self.productImage = aDecoder.decodeObject(forKey:"productImage") as! String?
         self.cellImage = aDecoder.decodeObject(forKey:"cellImage") as! String?
-        self.price = aDecoder.decodeObject(forKey: "price") as! Double?
+        self.price = aDecoder.decodeObject(forKey: "price") as! Int?
 
         
     }
@@ -37,6 +37,8 @@ class Product:  NSObject , NSCoding {
         aCoder.encode(self.cellImage,forKey:"cellImage")
         aCoder.encode(self.price,forKey:"price")
     }
+    
+    
     
     
 }
